add new measurements support
This enables kube-burner-ocp to define its own measurement
functionality (example, network-policy-measurement.go,
router-advertisement-measurement.go). This is required to avoid
using openshift CRD in kube-burner measurements code.

It defines exampleLatency which is currently working
and can be used as a reference for other measurements like
router-advertisement-measurement.go.

Now kube-burner-ocp prepares measurementFactoryMap with the core
measurements (like pod, service latency) and additional measurements
which is defined in kube-burner-ocp (like exampleLatency).

I have tested this code with only enabling podLatency and exampleLatency
in cluster-density-v2 workload

I will fine tune it once we are fine with this approach.

This PR depends on kube-burner PR
https://github.com/kube-burner/kube-burner/pull/841

Signed-off-by: venkataanil <vkommadi@redhat.com>