Fix BGP regression (#277)
* Fix BGP regression

BGP workload is crashing as explained in the bug #269
So reverting back the needed change.

Fixes #269

Signed-off-by: venkataanil <vkommadi@redhat.com>

* Fix measurements regression

RouteAdvertisement latency measurement code in kube-burner-ocp is
impacted by the kube-burner measurements refactoring PR #859
To address this, kube burner PR #925, now exports some variables
which will be used in kube burner ocp to fix the regression.

Fixes #278

Signed-off-by: venkataanil <vkommadi@redhat.com>

* using kube burner release v1.17.2

Signed-off-by: venkataanil <vkommadi@redhat.com>

---------

Signed-off-by: venkataanil <vkommadi@redhat.com>
Co-authored-by: Ygal Blum <ygal.blum@gmail.com>