[DNM] custom config for cluster density
kube-burner AI needs CI runs for cdv2 job with different config options.
We want cdv2 to create variable number of cluster resources in each test
run instead of predefined number of them.
Each iteration in the test run, choses a random number between the min
and max provided for the resource.

Please don't review or merge this change as it is a temporary change
for our CI runs.

Signed-off-by: venkataanil <vkommadi@redhat.com>